# openapi.yaml
openapi: 3.0.3
info:
  title: CivicOS API
  version: 1.0.0
servers:
  - url: https://civicos.onrender.com
paths:
...
  /api/notifications:
    get:
      summary: List notifications for current user
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
  /api/notifications/unread-count:
    get:
      summary: Get unread notifications count
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
  /api/notifications/read-all:
    patch:
      summary: Mark all notifications as read
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT